# docs/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Minimarket Plus API",
        "description": "API REST del backend de Minimarket Plus: productos, carrito, inventario, ventas y usuarios. Documentada con OpenAPI e hipermedios HATEOAS. Los recursos est\u00e1n protegidos por roles (CLIENTE, CAJERO, ADMIN).",
        "contact": {
            "name": "Equipo Minimarket Plus"
        },
        "license": {
            "name": "Uso acad\u00e9mico - Duoc UC"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8084",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
        "/api/usuarios/{id}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener usuario por ID",
                "operationId": "obtenerUsuarioPorId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityModelUsuario"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityModelUsuario"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar usuario",
                "operationId": "actualizarUsuario",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Usuario"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "404": {
                        "description": "Usuario no encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityModelUsuario"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Usuario actualizado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityModelUsuario"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Eliminar usuario",
                "operationId": "eliminarUsuario",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Usuario no encontrado"
                    },
                    "204": {
                        "description": "Usuario eliminado"
                    }
                }
            }
        },
        "/api/productos/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener producto por ID",
                "description": "Devuelve un producto espec\u00edfico con sus enlaces.",
                "operationId": "obtenerProductoPorId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Producto no encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityModelProducto"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Producto encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityModelProducto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar producto",
                "description": "Actualiza un producto existente. Requiere rol ADMIN.",
                "operationId": "actualizarProducto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Producto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "404": {
                        "description": "Producto no encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityModelProducto"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Producto actualizado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityModelProducto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar producto",
                "description": "Elimina un producto por ID. Requiere rol ADMIN.",
                "operationId": "eliminarProducto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "204": {
                        "description": "Producto eliminado"
                    }
                }
            }
        },
...
        "/api/usuarios": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Listar usuarios",
                "description": "Devuelve los usuarios registrados con enlaces HATEOAS.",
                "operationId": "listarUsuarios",
                "responses": {
                    "200": {
                        "description": "Listado obtenido correctamente",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CollectionModelEntityModelUsuario"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sin permisos (se requiere ADMIN)",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CollectionModelEntityModelUsuario"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Crear usuario",
                "description": "Registra un nuevo usuario. Requiere rol ADMIN.",
                "operationId": "guardarUsuario",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Usuario"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Usuario creado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityModelUsuario"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sin permisos (se requiere ADMIN)",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityModelUsuario"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Listar productos",
                "description": "Devuelve todos los productos con enlaces HATEOAS.",
                "operationId": "listarProductos",
                "responses": {
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CollectionModelEntityModelProducto"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Listado obtenido correctamente",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CollectionModelEntityModelProducto"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear producto",
                "description": "Registra un nuevo producto. Requiere rol ADMIN.",
                "operationId": "guardarProducto",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Producto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Producto creado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityModelProducto"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sin permisos (se requiere ADMIN)",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityModelProducto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Rol": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "type": "string"
                    }
                }
            },
            "Usuario": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string",
                        "writeOnly": true
                    },
                    "roles": {
                        "uniqueItems": true,
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Rol"
                        }
                    }
                }
            },
            "EntityModelUsuario": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string",
                        "writeOnly": true
                    },
                    "roles": {
                        "uniqueItems": true,
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Rol"
                        }
                    },
                    "_links": {
                        "$ref": "#/components/schemas/Links"
                    }
                }
            },
            "Links": {
                "type": "object",
                "additionalProperties": {
                    "$ref": "#/components/schemas/Link"
                }
            },
            "Categoria": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "type": "string"
                    }
                }
            },
            "Producto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "precio": {
                        "type": "number",
                        "format": "double"
                    },
                    "stock": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "categoria": {
                        "$ref": "#/components/schemas/Categoria"
                    }
                }
            },
            "EntityModelProducto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "precio": {
                        "type": "number",
                        "format": "double"
                    },
                    "stock": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "categoria": {
                        "$ref": "#/components/schemas/Categoria"
                    },
                    "_links": {
                        "$ref": "#/components/schemas/Links"
                    }
                }
            },
...
            "CollectionModelEntityModelUsuario": {
                "type": "object",
                "properties": {
                    "_embedded": {
                        "type": "object",
                        "properties": {
                            "usuarioList": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/EntityModelUsuario"
                                }
                            }
                        }
                    },
                    "_links": {
                        "$ref": "#/components/schemas/Links"
                    }
                }
            },
            "CollectionModelEntityModelProducto": {
                "type": "object",
                "properties": {
                    "_embedded": {
                        "type": "object",
                        "properties": {
                            "productoList": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/EntityModelProducto"
                                }
                            }
                        }
                    },
                    "_links": {
                        "$ref": "#/components/schemas/Links"
                    }
                }
            },
...
            "Link": {
                "type": "object",
                "properties": {
                    "href": {
                        "type": "string"
                    },
                    "hreflang": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "deprecation": {
                        "type": "string"
                    },
                    "profile": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "templated": {
                        "type": "boolean"
                    }
                }
            }
        }
    }
}